Refactor status options and update imports for consistency
