+++ src/features/classDetail/constant/statusEnum.js
@@ -1,6 +1,5 @@
-export const StatusEnum = Object.freeze({
-  ALL: "",
-  COMPLETE: "COMPLETE",
-  ONGOING: "ON_GOING",
-  NOTSTARTED: "NOT_STARTED",
-});
+export const statusOptions = {
+  NOT_STARTED: { label: "Not started", bg: "#E5E7EB", text: "#374151" },
+  ON_GOING: { label: "Ongoing", bg: "#E1E8FF", text: "#1C3FB7" },
+  COMPLETE: { label: "Complete", bg: "#DAF8E6", text: "#1A8245" },
+};
