+++ src/features/classDetail/validate.js
@@ -28,9 +28,3 @@ export const sessionSchema = yup.object().shape({
       return isStartDateValid(value?.[0]);
     }),
 });
-
-export const statusOptions = {
-  NOT_STARTED: { label: "Not started", bg: "#E5E7EB", text: "#374151" },
-  ON_GOING: { label: "Ongoing", bg: "#E1E8FF", text: "#1C3FB7" }, // xanh dương
-  COMPLETED: { label: "Completed", bg: "#DAF8E6", text: "#1A8245" }, // xanh lá
-};
